feat: update tests
